% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Split a vector into n chunks}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{a vector}

\item{n}{number of chunks}
}
\description{
Copied from the \code{stackoverflow} package.
}
\examples{
 chunk(1:30, 6)
}
\references{
\url{http://stackoverflow.com/questions/3318333/split-a-vector-into-chunks-in-r}
}
\author{
\href{http://stackoverflow.com/users/1563634/mathheadinclouds}{mathheadinclouds},
\href{http://stackoverflow.com/users/1737569/dis-shishkov}{Dis Shishkov}
}
