% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_and_fit.R
\name{dgp}
\alias{dgp}
\title{Data Generating Process}
\usage{
dgp(spec, param, n, R, seed, burn.in = 25000L, nthreads = 4L, ...)
}
\arguments{
\item{spec}{specification object create by \code{SLSAGARCH2::specification}}

\item{param}{true parameters of the DGP}

\item{n}{sample size}

\item{R}{number of Monte Carlo replications}

\item{seed}{the seed for the rng}

\item{burn.in}{the burnin period}

\item{nthreads}{number of threads to use for parallel simulation}
}
\value{
An n x R matrix of simulated observations.
}
\description{
Generates observations from the specified model with
hyperbolic secant distributed innovations.
}
