% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{estimate_scenario}
\alias{estimate_scenario}
\title{Estimate Scenario}
\usage{
estimate_scenario(
  spec = SLSAGARCH::specification(type = "S"),
  se_type = "QMLE",
  tiktak = TRUE,
  tiktak_opts = list(n = 250L, N = 10L, thin = 5L, reltol = 0.001, lead_solver_opts =
    SLSAGARCH::nloptr_slsqp_options(maxtime = 0.35), lag_solver_opts =
    SLSAGARCH::nloptr_slsqp_options(maxtime = 0.35/2), a = 2, b = 2.5, m = 25L, verbose =
    FALSE),
  strategy = "multisession",
  scenario = 1L,
  ...
)
}
\arguments{
\item{spec}{model specification}

\item{se_type}{type of standard errors, default is "QLIKE"}

\item{tiktak}{boolean indicating whether to use \code{SLSAGARCH::tiktak}}

\item{tiktak_opts}{list of options to pass to \code{SLSAGARCH::tiktak}}

\item{strategy}{strategy for parallel execution, default is "multisession"}

\item{scenario}{scenario identifier.}

\item{...}{additional parameters to pass to \code{fit_many}}
}
\value{
A dataframe
}
\description{
Reads simulated data for a scenario and fits the
SLS-AGARCH(1, 1) model to each realisation. Returns
a data frame suitable for use with \code{rsimsum}
}
