% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{fit}
\alias{fit}
\title{Fit Specified Model to Simulated Data}
\usage{
fit(spec, param, y, se_type = "QMLE", ...)
}
\arguments{
\item{spec}{specification object create by \code{SLSAGARCH::specification}}

\item{param}{true parameters of the DGP}

\item{y}{simulated observations}

\item{se_type}{method for calculating the standard errors}

\item{...}{additional arguments passed to \code{SLSAGARCH::estimate}}
}
\value{
A tibble with estimates
}
\description{
This is a simple wrapper around \code{SLSAGARCH::estimate} that
produces a tibble suitably for further analysis.
}
\examples{
# Create the model specification
spec <- specification(type = "S")
# True parameter values
param0 <- c(h0 = 1, omega = 0.1, alpha = 0.075, beta = 0.85, lambda = 1, rho = 0.75)
#   Set the initial conditional volatility equal to the unconditional volatility.
param0[1] <- SLSAGARCH:::unconditional.specification(spec, param0)
# Simulate 10000 series, each of length 500
Y <- dgp(spec, param0, n = 500, R = 10000, seed = 20240127, nthreads = 10L)
fit(spec, param0, Y[, 1])
}
