% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Simulation Parameters}
\format{
A 9x6 data frame with a column for each parameter.
}
\source{
Own calculations.
}
\usage{
parameters
}
\description{
Parameters used to simulate data.
The value of omega is selected to ensure that the unconditional volatility
equals one.
}
\keyword{datasets}
