% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prhs}
\alias{prhs}
\title{Simulate From a Hyperbolic Secant Distribution in Parallel}
\usage{
prhs(n, m, seed, numThreads = -1L)
}
\arguments{
\item{n}{Number of observations to generate each replication}

\item{m}{Number of replications}

\item{seed}{Seed for the rng}

\item{numThreads}{Number of threads for running code in parallel}
}
\description{
Simulates \code{n} observations from \code{m} independent hyperbolic secant
distributed random variables.
}
\details{
Note that \code{seed} works in cojunction with \code{numThreads}. Therefore,
setting \code{seed = 42} and \code{numThreads = 2} will produce differents
results compared to setting \code{seed = 42} and \code{numThreads = 4}.
}
