% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario_01}
\alias{scenario_01}
\title{Scenario 1}
\format{
A list of matrices that contain the simulated observations.
}
\source{
Own calculations.
}
\usage{
scenario_01
}
\description{
Data for the case where \eqn{\lambda = 0,} \eqn{\rho = 0} and
\eqn{n \in \{250, 500, 750, 1000, 1250, 2500 \},} i.e.,
the 1st row of \code{parameters}.
}
\keyword{datasets}
