% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario_05}
\alias{scenario_05}
\title{Scenario 5}
\format{
A list of matrices that contain the simulated observations.
}
\source{
Own calculations.
}
\usage{
scenario_05
}
\description{
Data for the case where \eqn{\lambda = 0.5,} \eqn{\rho = 0.25} and
\eqn{n \in \{250, 500, 750, 1000, 1250, 2500 \},} i.e.,
the 5th row of \code{parameters}.
}
\keyword{datasets}
