% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario_09}
\alias{scenario_09}
\title{Scenario 9}
\format{
A list of matrices that contain the simulated observations.
}
\source{
Own calculations.
}
\usage{
scenario_09
}
\description{
Data for the case where \eqn{\lambda = 1,} \eqn{\rho = 0.75} and
\eqn{n \in \{250, 500, 750, 1000, 1250, 2500 \},} i.e.,
the 9th row of \code{parameters}.
}
\keyword{datasets}
