% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scenario_files}
\alias{scenario_files}
\title{Scenarios}
\usage{
scenario_files(path = NULL)
}
\arguments{
\item{path}{file name}
}
\value{
either a vector of file names or the full path to a particular file
}
\description{
Lists files that store simulated data.
}
\examples{
scenario_files()
scenario_files(path = "scenario_01.fst")
}
