\name{SimStudySLSAGARCH-package}
\alias{SimStudySLSAGARCH-package}
\alias{SimStudySLSAGARCH}
\docType{package}
\title{\packageTitle{SimStudySLSAGARCH}}
\description{\packageDescription{SimStudySLSAGARCH}}
\section{Package Content}{\packageIndices{SimStudySLSAGARCH}}
\author{\packageAuthor{SimStudySLSAGARCH}}
\section{Maintainer}{\packageMaintainer{SimStudySLSAGARCH}}
\keyword{package}
