% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sls_sech_spec}
\alias{sls_sech_spec}
\title{Hyperbolic Secant SLSAGARCH(1, 1) Model Specification}
\format{
An object of type specification, from the SLS-AGARCH package.
}
\source{
Own calculations.
}
\usage{
sls_sech_spec
}
\description{
The specification object for an SLSAGARCH(1, 1) model with
hyperbolic secant distributed innovations. Estimation is
subject to the restriction that the unconditional variance
exists and the model is invertible.
}
\keyword{datasets}
