% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{estimate_scenario}
\alias{estimate_scenario}
\title{Estimate Scenario}
\usage{
estimate_scenario(
  spec = SLSAGARCH2:::specification(type = "S", distribution = "sech"),
  se_type = "QMLE",
  strategy = "multisession",
  scenario = 1L,
  ...
)
}
\arguments{
\item{spec}{model specification}

\item{se_type}{type of standard errors, default is "QLIKE"}

\item{strategy}{strategy for parallel execution, default is "multisession"}

\item{scenario}{scenario identifier.}

\item{...}{additional parameters to pass to \code{fit_many}}
}
\value{
A dataframe
}
\description{
Reads simulated data for a scenario and fits the
SLS-AGARCH(1, 1) model to each realisation. Returns
a data frame suitable for use with \code{rsimsum}
}
