% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{fit}
\alias{fit}
\title{Fit Specified Model to Simulated Data}
\usage{
fit(spec, param, y, se_type = "QMLE", ...)
}
\arguments{
\item{spec}{specification object create by \code{SLSAGARCH::specification}}

\item{param}{true parameters of the DGP}

\item{y}{simulated observations}

\item{se_type}{method for calculating the standard errors}

\item{...}{additional arguments passed to \code{SLSAGARCH::estimate}}
}
\value{
A tibble with estimates
}
\description{
This is a simple wrapper around \code{SLSAGARCH::estimate} that
produces a tibble suitably for further analysis.
}
