% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{fit_many}
\alias{fit_many}
\title{Fit Many}
\usage{
fit_many(
  spec,
  param,
  ymat,
  se_type = "QMLE",
  label = sim_lab(n = nrow(ymat), lambda = param["lambda"], rho = param["rho"]),
  strategy = "sequential",
  nchunks = 100,
  .errorhandling = "stop",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{spec}{specification object create by \code{SLSAGARCH::specification}}

\item{param}{true parameters of the DGP}

\item{ymat}{matrix of simulated observations}

\item{se_type}{method for calculating the standard errors}

\item{label}{a label used to identify the simulation}

\item{strategy}{evaluation function (or name of it) to use for resolving a future. If NULL, then the current strategy is returned.}

\item{nchunks}{the number of realisations to pass to each thread}

\item{.errorhandling}{specifies how a task evaluation error should be handled. If the value is "stop", then execution will be stopped via the stop function if an error occurs. If the value is "remove", the result for that task will not be returned, or passed to the .combine function. If it is "pass", then the error object generated by task evaluation will be included with the rest of the results. It is assumed that the combine function (if specified) will be able to deal with the error object. The default value is "stop".}

\item{verbose}{logical flag enabling verbose messages}

\item{...}{}
}
\value{
a dataframe
}
\description{
Fit Many
}
\details{
This function creates a \code{progressor} that signals progress updates. The user
can run the function inside of a \code{with_progress} function to show a progress
bar.
}
\examples{
# Create the model specification
spec <- specification(type = "S")
# True parameter values
param0 <- c(h0 = 1, omega = 0.1, alpha = 0.075, beta = 0.85, lambda = 1, rho = 0.75)
#   Set the initial conditional volatility equal to the unconditional volatility.
param0[1] <- SLSAGARCH:::unconditional.specification(spec, param0)
# Simulate 10000 series, each of length 500
Y <- dgp(spec, param0, n = 500, R = 10000, seed = 20240127, nthreads = 10L)
fit_many(spec, param0, Y, strategy = "multisession", nchunks = 100)
# if you want to see a progress bar...
progressr::with_progress(fit_many(spec, param0, Y, strategy = "multisession", nchunks = 100))
}
